[WIP] finish unit tests and refactoring
