+++ tests/helper.py
@@ -24,6 +24,9 @@ from dataclasses import dataclass
 import os
 import pathlib
 import time
+from typing import Any
+import pytest
+import numpy
 
 import ansys.dpf.core as dpf
 from ansys.dpf.core import DataSources, Field, MeshedRegion, Operator
@@ -156,3 +159,54 @@ def get_dummy_data_files(distributed: bool = False):
         composite={"shell": CompositeDefinitionFiles(definition=h5_path)},
         engineering_data=material_path,
     )
+
+
+def compare_sampling_point_results(
+        sample: dict[Any, Any],
+        reference: dict[Any, Any],
+        with_polar_properties: bool
+) -> None:
+    assert sample["element_label"] == reference["element_label"]
+    assert sample["unit_system"] == reference["unit_system"]
+
+    assert sample["layup"]["num_analysis_plies"] == reference["layup"]["num_analysis_plies"]
+    assert sample["layup"]["offset"] == reference["layup"]["offset"]
+    if not with_polar_properties:
+        assert sample["layup"]["polar_properties"] is None
+    else:
+        raise RuntimeError("Float comparison of polar properties is not implemented.")
+
+    assert len(sample["layup"]["analysis_plies"]) == len(reference["layup"]["analysis_plies"])
+    for ap_index, ap in enumerate(sample["layup"]["analysis_plies"]):
+        ref_ap = reference["layup"]["analysis_plies"][ap_index]
+        assert ap.keys() == ref_ap.keys()
+        for key in ap.keys():
+            if key in ["angle", "thickness"]:
+                assert ap[key] == pytest.approx(ref_ap[key], abs=1e-8, rel=1e-6)
+            else:
+                assert ap[key] == ref_ap[key]
+
+    result_keys = ['failures', 'strains', 'stresses', 'offsets']
+    reference_results = reference['results']
+    sample_results = sample['results']
+    for key in result_keys:
+        if key == "offsets":
+            numpy.testing.assert_allclose(
+                sample_results[key],
+                reference_results[key],
+                rtol=1e-6,
+                atol=1e-8
+            )
+        else:
+            assert sample_results[key].keys() == reference_results[key].keys()
+            for component in sample_results[key].keys():
+                if component == 'failure_modes':
+                    # compare list of strings
+                    assert sample_results[key][component] == reference_results[key][component]
+                else:
+                    numpy.testing.assert_allclose(
+                        sample_results[key][component],
+                        reference_results[key][component],
+                        rtol=1e-6,
+                        atol=1e-8
+                    )
