+++ tests/sampling_point_solid_stack_test.py
@@ -19,7 +19,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
+import json
 import os
 import pathlib
 
@@ -41,13 +41,42 @@ from ansys.dpf.composites.failure_criteria import (
     MaxStressCriterion,
     PuckCriterion,
     FaceSheetWrinklingCriterion,
-    CoreFailureCriterion
+    CoreFailureCriterion,
 )
 from ansys.dpf.composites.result_definition import FailureMeasureEnum
 from ansys.dpf.composites.sampling_point_types import FailureResult
 
-from .helper import get_basic_shell_files
+from .helper import compare_sampling_point_results
 
+TEST_DATA_ROOT_DIR = pathlib.Path(__file__).parent / "data" / "solid_model"
+
+def get_file_paths() -> ContinuousFiberCompositesFiles:
+    rst_paths = [os.path.join(TEST_DATA_ROOT_DIR, "file.rst")]
+    h5_path = os.path.join(TEST_DATA_ROOT_DIR, "ACPSolidModel_SolidModel.1.h5")
+    material_path = os.path.join(TEST_DATA_ROOT_DIR, "material.engd")
+
+    files = ContinuousFiberCompositesFiles(
+        rst=rst_paths,
+        composite={"shell": CompositeDefinitionFiles(definition=h5_path)},
+        engineering_data=material_path,
+    )
+    return files
+
+def get_fc() -> CombinedFailureCriterion:
+    cfc = CombinedFailureCriterion("combined")
+    cfc.insert(MaxStrainCriterion(e3 = True, e13 = True, e23 = True))
+    cfc.insert(MaxStressCriterion(s3 = True, s13 = True, s23 = True))
+    cfc.insert(PuckCriterion(pd = True, dim = 3))
+    cfc.insert(FaceSheetWrinklingCriterion())
+    cfc.insert(CoreFailureCriterion(include_ins = True))
+    return cfc
+
+
+def compare_results(current, element_id) -> None:
+    with open(os.path.join(TEST_DATA_ROOT_DIR, f"ref_{element_id}.txt")) as f:
+        reference = json.loads(json.load(f))
+
+    compare_sampling_point_results(reference, current, with_polar_properties=False)
 
 def test_sampling_point_solid_stack(dpf_server):
     """
@@ -55,30 +84,11 @@ def test_sampling_point_solid_stack(dpf_server):
 
     The solid stack contains only layered elements.
     """
-
-    TEST_DATA_ROOT_DIR = r"D:\tmp\small_solid_model_files\dp0\SYS\MECH"
-    rst_paths = [os.path.join(TEST_DATA_ROOT_DIR, "file.rst")]
-    h5_path = os.path.join(TEST_DATA_ROOT_DIR, "Setup", "ACPSolidModel_SolidModel.1.h5")
-    material_path = os.path.join(TEST_DATA_ROOT_DIR, "Setup", "material.engd")
-
-    files = ContinuousFiberCompositesFiles(
-        rst=rst_paths,
-        composite={"shell": CompositeDefinitionFiles(definition=h5_path)},
-        engineering_data=material_path,
-    )
-
-    composite_model = CompositeModel(files, server=dpf_server)
-
-    cfc = CombinedFailureCriterion("combined")
-    cfc.insert(MaxStrainCriterion())
-    cfc.insert(MaxStressCriterion())
-    cfc.insert(PuckCriterion())
-    cfc.insert(FaceSheetWrinklingCriterion())
-    cfc.insert(CoreFailureCriterion())
+    composite_model = CompositeModel(get_file_paths(), server=dpf_server)
 
     # Element with a standard solid stack (hex only)
     sampling_point = composite_model.get_sampling_point(
-        combined_criterion=cfc,
+        combined_criterion=get_fc(),
         element_id=110,
         time=1
     )
@@ -90,11 +100,8 @@ def test_sampling_point_solid_stack(dpf_server):
         core_scale_factor=0.2,
         show_failure_modes=True,
     )
-    plot.figure.show()
-    res = sampling_point.results
-    print(res[0]['results']['strains']['e12'])
-
-
+    # plot.figure.show()
+    compare_results(sampling_point.results[0], 110)
 
 def test_sampling_point_solid_stack_with_dropoffs(dpf_server):
     """
@@ -102,29 +109,13 @@ def test_sampling_point_solid_stack_with_dropoffs(dpf_server):
 
     The drop-off elements need special handling to extract the results and layup information.
     """
-    TEST_DATA_ROOT_DIR = r"D:\tmp\small_solid_model_files\dp0\SYS\MECH"
-    rst_paths = [os.path.join(TEST_DATA_ROOT_DIR, "file.rst")]
-    h5_path = os.path.join(TEST_DATA_ROOT_DIR, "Setup", "ACPSolidModel_SolidModel.1.h5")
-    material_path = os.path.join(TEST_DATA_ROOT_DIR, "Setup", "material.engd")
+    composite_model = CompositeModel(get_file_paths(), server=dpf_server)
 
-    files = ContinuousFiberCompositesFiles(
-        rst=rst_paths,
-        composite={"shell": CompositeDefinitionFiles(definition=h5_path)},
-        engineering_data=material_path,
-    )
-
-    composite_model = CompositeModel(files, server=dpf_server)
-
-    cfc = CombinedFailureCriterion("combined")
-    cfc.insert(MaxStrainCriterion())
-    cfc.insert(MaxStressCriterion())
-    cfc.insert(PuckCriterion())
-    cfc.insert(FaceSheetWrinklingCriterion())
-    cfc.insert(CoreFailureCriterion())
-
+    # one drop-off with one element only
+    # one drop-off which is split into two tetra elements
     for element_id in [113, 154]:
         sampling_point = composite_model.get_sampling_point(
-            combined_criterion=cfc,
+            combined_criterion=get_fc(),
             element_id=element_id,
             time=1
         )
@@ -136,9 +127,8 @@ def test_sampling_point_solid_stack_with_dropoffs(dpf_server):
             core_scale_factor=0.2,
             show_failure_modes=True,
         )
-        plot.figure.show()
-        res = sampling_point.results
-        print(res[0]['results']['strains']['e12'])
+        # plot.figure.show()
+        compare_results(sampling_point.results[0], element_id)
 
 
 def test_sampling_point_solid_stack_with_cutoffs(dpf_server):
@@ -147,29 +137,11 @@ def test_sampling_point_solid_stack_with_cutoffs(dpf_server):
 
     The cut-off elements need special handling to extract the results and layup information.
     """
-    TEST_DATA_ROOT_DIR = r"D:\tmp\small_solid_model_files\dp0\SYS\MECH"
-    rst_paths = [os.path.join(TEST_DATA_ROOT_DIR, "file.rst")]
-    h5_path = os.path.join(TEST_DATA_ROOT_DIR, "Setup", "ACPSolidModel_SolidModel.1.h5")
-    material_path = os.path.join(TEST_DATA_ROOT_DIR, "Setup", "material.engd")
-
-    files = ContinuousFiberCompositesFiles(
-        rst=rst_paths,
-        composite={"shell": CompositeDefinitionFiles(definition=h5_path)},
-        engineering_data=material_path,
-    )
-
-    composite_model = CompositeModel(files, server=dpf_server)
-
-    cfc = CombinedFailureCriterion("combined")
-    cfc.insert(MaxStrainCriterion())
-    cfc.insert(MaxStressCriterion())
-    cfc.insert(PuckCriterion())
-    cfc.insert(FaceSheetWrinklingCriterion())
-    cfc.insert(CoreFailureCriterion())
+    composite_model = CompositeModel(get_file_paths(), server=dpf_server)
 
     # Element of a solid stack with cut-off elements
     sampling_point = composite_model.get_sampling_point(
-        combined_criterion=cfc,
+        combined_criterion=get_fc(),
         element_id=219,
         time=1
     )
@@ -181,6 +153,5 @@ def test_sampling_point_solid_stack_with_cutoffs(dpf_server):
         core_scale_factor=0.2,
         show_failure_modes=True,
     )
-    plot.figure.show()
-    res = sampling_point.results
-    print(res[0]['results']['strains']['e12'])
+    # plot.figure.show()
+    compare_results(sampling_point.results[0], 219)
